Prune repository to essential MSc coursework files
- Inline functions from src/data.py into train.py, eval.py, and train_ann_baseline.py
- Remove entire src/ directory, Makefile, and pyproject.toml
- Update CI workflow to syntax-check only (no training or linting)
- Add missing dependencies to requirements.txt (pyyaml, pandas, opencv-python)
- Update README with 'Files included' section documenting kept files
- Ensure results/ and report_assets/ directories exist with .gitkeep files
- Maintain Windows compatibility with hardcoded C:/PlantVillage path
- All syntax checks pass successfully
