+++ .github/workflows/ci.yml
@@ -21,31 +21,14 @@ jobs:
       with:
         python-version: ${{ matrix.python-version }}
     
-    - name: Cache pip packages
-      uses: actions/cache@v4
-      with:
-        path: ~/.cache/pip
-        key: ${{ runner.os }}-pip-${{ hashFiles('**/requirements.txt') }}
-        restore-keys: |
-          ${{ runner.os }}-pip-
-    
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
         pip install -r requirements.txt
-        pip install ruff black
     
-    - name: Lint with ruff
+    - name: Import check
       run: |
-        ruff check src/
-    
-    - name: Check formatting with black
-      run: |
-        black --check src/
-    
-    - name: Run synthetic smoke test
-      run: |
-        make smoke
+        python -m py_compile train.py eval.py visualize_results.py gradcam.py cv_viz.py ann_dataset.py train_ann_baseline.py
     
     - name: Upload artifacts
       uses: actions/upload-artifact@v4
@@ -53,6 +36,6 @@ jobs:
       with:
         name: test-artifacts-py${{ matrix.python-version }}
         path: |
+          results/
           report_assets/
-          checkpoints/
         retention-days: 7
