--- src/__init__.py
@@ -1,3 +1,0 @@
-"""PlantVillage disease classification package."""
-
-__version__ = "0.1.0"
