--- src/models/__init__.py
@@ -1,6 +1,0 @@
-"""Model definitions for plant disease classification."""
-
-from .ann import HOGClassifier
-from .cnn import EfficientNetClassifier, MobileNetClassifier
-
-__all__ = ["EfficientNetClassifier", "MobileNetClassifier", "HOGClassifier"]
